introduce viewModel for text selection/mark
